+++ src/annie.js
@@ -153,7 +153,7 @@ class Annie extends Discord.Client {
      * @returns {void}
      */
     prepareLogin() {
-    process.on(`unhandledRejection`, err => logger.warn(`Catched rejection > ${err.message}`))
+    process.on(`unhandledRejection`, err => logger.warn(`Catched rejection > ${err.stack}`))
         try {
             this._initializingDatabase()
             this._initializingCommands()
@@ -167,22 +167,23 @@ class Annie extends Discord.Client {
     }
 
     /**
-     * Registering configuration nodes for each guild
+     * Registering configuration nodes for each guild or single guild by specifying it in the parameter.
+     * @param {string} [guildId=null] Specify target guild id for single guild register.
      * @author klerikdust
-     * @returns {void}
+     * @return {void}
      */
-    async registerGuildConfigurations() {
+    async registerGuildConfigurations(guildId=null) {
         const initTime = process.hrtime()
-        const configClass = new customConfig(this)
         const registeredGuildConfigurations = await this.db.getAllGuildsConfigurations()
-        const getGuilds = this.guilds.cache.map(guild => guild.id)
+        //  If prompted to register only single guild, then use single-element array.
+        const getGuilds = guildId ? [guildId] : this.guilds.cache.map(guild => guild.id)
         for (let i=0; i<getGuilds.length; i++) {
             let guild = this.guilds.cache.get(getGuilds[i])
             let existingGuildConfigs = registeredGuildConfigurations.filter(node => node.guild_id === guild.id)
             guild.configs = new Map()
             //  Iterating over all the available configurations
-            for (let x=0; x<configClass.availableConfigurations.length; x++) {
-                let cfg = configClass.availableConfigurations[x]
+            for (let x=0; x<customConfig.availableConfigurations.length; x++) {
+                let cfg = customConfig.availableConfigurations[x]
                 //  Register existing configs into guild's nodes if available
                 if (existingGuildConfigs.length > 0) {
                     const matchConfigCode = existingGuildConfigs.filter(node => node.config_code.toUpperCase() === cfg.name)[0]
@@ -195,8 +196,8 @@ class Annie extends Discord.Client {
                 }
                 guild.configs.set(cfg.name, cfg)
             }
-        }
-        logger.info(`[SHARD_ID:${this.shard.ids[0]}@GUILD_CONF] confs from ${getGuilds.length} guilds have been registered (${getBenchmark(initTime)})`)
+        } 
+        this.logger.info(`[SHARD_ID:${this.shard.ids[0]}@GUILD_CONF] confs from ${getGuilds.length} guilds have been registered (${getBenchmark(initTime)})`)
     }
 
     /**