+++ src/libs/exp.js
@@ -63,6 +63,7 @@ class Experience {
     execute(expToBeAdded=this.defaultGain) {
     	this.client.db.getUserExp(this.user.id, this.guild.id)
         .then(exp => {
+            if (!exp) return
             const prevExp = this.xpFormula(exp.current_exp)
             const newExp = this.xpFormula(exp.current_exp + expToBeAdded)
             //  Send level up message if new level is higher than previous level
@@ -88,23 +89,30 @@ class Experience {
         if (registeredRanks.length <= 0) return
         const userRankLevel = closestBelow(registeredRanks.map(r => r.LEVEL), level) 
         //  Early exit on unranked level. Also remove all the existing rank role.
-        if (!isFinite(userRankLevel)) return this.user.roles.remove(registeredRanks.map(r => r.ROLE))
-        .catch(e => this.client.logger.warn(`${this.instanceId} <FAIL> role clear > ${e.message}`))
+        if (!isFinite(userRankLevel)) return
         const expectedTargetRole = registeredRanks.find(r => parseInt(r.LEVEL) === userRankLevel)
         if (!expectedTargetRole) return
         //  Ensure that role is exist in the server/guild.
         const userRankRole = this.guild.roles.cache.get(expectedTargetRole.ROLE)
         if (!userRankRole) return
-        //  Assign new rank role if target level is not -infinity (due to below threshold).
-        if (isFinite(userRankLevel)) this.user.roles.add(userRankRole)
-        .catch(e => this.client.logger.warn(`${this.instanceId} <FAIL> role assign > ${e.message}`))
+        //  Assign new rank role 
+        try {
+            await this.user.roles.add(userRankRole)
+        }
+        catch(e) {
+            this.client.logger.warn(`${this.instanceId} <FAIL> role assign > ${e.message}`)
+        }
         if (this.guild.configs.get(`RANKS_STACK`).value) return
         //  Remove non-current rank roles if RANKS_STACK is disabled
         const nonCurrentRankRoles = registeredRanks
         .filter(r => r.ROLE !== userRankRole.id)
         .map(r => r.ROLE)
-        this.user.roles.remove(nonCurrentRankRoles)
-        .catch(e => this.client.logger.warn(`${this.instanceId} <FAIL> role remove > ${e.message}`))
+        try {
+            this.user.roles.remove(nonCurrentRankRoles)
+        }
+        catch(e) {
+            this.client.logger.warn(`${this.instanceId} <FAIL> role remove > ${e.message}`)
+        }
 	}
 
     /**