+++ src/utils/loadAsset.js
@@ -15,6 +15,10 @@ const loadAsset = async (id=``, assetsPath=`./src/assets`) => {
 		if (fileWithoutFormat === id) return ultimateFile = file 
 	})
 	if (!ultimateFile) {
+        //  Replace path with default one
+        assetsPath = `./src/assets`
+        //  Refetch
+	    allFiles = fs.readdirSync(assetsPath)
 		allFiles.forEach((f) => {
 			if (f.includes(`defaultcover1`)) {
 				let filePath = `./${f.replace(/\\/g, `/`)}`